# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Cetacean API
  version: "1.0"
  description: |
    Observability and management API for Docker Swarm Mode clusters.

    Uses content negotiation — same URLs serve JSON, HTML (SPA), or SSE based on
    the Accept header. Append `.json` or `.html` to any path to force a content type.

    All collection responses include JSON-LD `@context` and `@type` fields.
    Detail responses include `@context`, `@id`, and `@type`.
    Error responses follow RFC 9457 (Problem Details).

    Every response (except `/-/` meta endpoints) includes:
    - `Link: </api>; rel="service-desc"` — pointer to this spec
    - `Link: </api/context.jsonld>; rel="describedby"` — JSON-LD context
    - `Vary: Accept` — content negotiation indicator
servers:
  - url: /
security:
  - cookieAuth: []
  - bearerAuth: []
  - {} # Allow unauthenticated (none mode)
...
paths:
...
  /nodes:
    get:
      operationId: listNodes
      tags: [Nodes]
      summary: List nodes
      parameters:
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/filter"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - name: sort
          in: query
          schema:
            type: string
            enum: [hostname, role, status, availability]
        - $ref: "#/components/parameters/dir"
      responses:
        "200":
          description: Paginated node list
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Link:
              $ref: "#/components/headers/Link"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeCollection"
        "304":
          $ref: "#/components/responses/NotModified"
        "400":
          $ref: "#/components/responses/BadRequest"
  /nodes/{id}:
    get:
      operationId: getNode
      tags: [Nodes]
      summary: Node detail
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "200":
          description: Node detail
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/DetailEnvelope"
                  - type: object
                    properties:
                      node:
                        type: object
                        description: Docker swarm.Node object
        "304":
          $ref: "#/components/responses/NotModified"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: removeNode
      tags: [Nodes]
      summary: Remove a node
      description: Removes a node from the swarm. The node must be drained or down. Requires operations level 3.
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "204":
          description: Node removed
        "403":
          description: Operations level too low
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /services:
    get:
      operationId: listServices
      tags: [Services]
      summary: List services
      parameters:
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/filter"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, mode]
        - $ref: "#/components/parameters/dir"
      responses:
        "200":
          description: Paginated service list
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Link:
              $ref: "#/components/headers/Link"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceCollection"
        "304":
          $ref: "#/components/responses/NotModified"
        "400":
          $ref: "#/components/responses/BadRequest"
  /services/{id}:
    get:
      operationId: getService
      tags: [Services]
      summary: Service detail
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "200":
          description: Service detail
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/DetailEnvelope"
                  - type: object
                    properties:
                      service:
                        type: object
                        description: Docker swarm.Service object
                      integrations:
                        type: array
                        description: Detected third-party integrations (e.g. Traefik)
                        items:
                          $ref: "#/components/schemas/Integration"
        "304":
          $ref: "#/components/responses/NotModified"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: removeService
      tags: [Services]
      summary: Remove a service
      description: Removes a service from the swarm. Requires operations level 3.
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "204":
          description: Service removed
        "403":
          description: Operations level too low
        "404":
          $ref: "#/components/responses/NotFound"
...
  /tasks:
    get:
      operationId: listTasks
      tags: [Tasks]
      summary: List tasks
      description: |
        Tasks are returned as EnrichedTask objects with ServiceName and
        NodeHostname populated from cache cross-references.
      parameters:
        - $ref: "#/components/parameters/filter"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - name: sort
          in: query
          schema:
            type: string
            enum: [state, service, node]
        - $ref: "#/components/parameters/dir"
      responses:
        "200":
          description: Paginated task list
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Link:
              $ref: "#/components/headers/Link"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskCollection"
        "304":
          $ref: "#/components/responses/NotModified"
        "400":
          $ref: "#/components/responses/BadRequest"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [Tasks]
      summary: Task detail
      description: |
        Returns the enriched task with cross-references to its service and node.
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "200":
          description: Task detail with service and node refs
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/DetailEnvelope"
                  - type: object
                    properties:
                      task:
                        $ref: "#/components/schemas/EnrichedTask"
                      service:
                        type: object
                        description: Cross-reference to the parent service
                        properties:
                          "@id":
                            type: string
                          name:
                            type: string
                      node:
                        type: object
                        description: Cross-reference to the assigned node
                        properties:
                          "@id":
                            type: string
                          hostname:
                            type: string
        "304":
          $ref: "#/components/responses/NotModified"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: removeTask
      tags: [Tasks]
      summary: Force-remove a task
      description: |
        Forces removal of a task (equivalent to `docker rm -f` on the underlying container).
        Only works on tasks whose containers are still accessible on the node.
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "204":
          description: Task removed
        "404":
          $ref: "#/components/responses/NotFound"
...
  /secrets:
    get:
      operationId: listSecrets
      tags: [Secrets]
      summary: List secrets
      description: Secret data is always redacted (cleared before response).
      parameters:
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/filter"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, created, updated]
        - $ref: "#/components/parameters/dir"
      responses:
        "200":
          description: Paginated secret list
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Link:
              $ref: "#/components/headers/Link"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecretCollection"
        "304":
          $ref: "#/components/responses/NotModified"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      operationId: createSecret
      tags: [Secrets]
      summary: Create a secret
      description: Creates a new secret in the swarm. Requires operations level 2.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, data]
              properties:
                name:
                  type: string
                  description: Secret name
                data:
                  type: string
                  format: byte
                  description: Secret data (base64-encoded)
      responses:
        "201":
          description: Secret created
          headers:
            Location:
              schema:
                type: string
              description: Path to the created secret
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/DetailEnvelope"
                  - type: object
                    properties:
                      secret:
                        type: object
                        description: Docker swarm.Secret object (data is always cleared)
                      services:
                        type: array
                        items:
                          $ref: "#/components/schemas/ServiceRef"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Operations level too low
        "409":
          description: Secret name conflict
  /secrets/{id}:
    get:
      operationId: getSecret
      tags: [Secrets]
      summary: Secret detail
      description: |
        Returns the secret metadata (data always redacted) and
        a list of services using this secret.
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "200":
          description: Secret detail with cross-referenced services
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/DetailEnvelope"
                  - type: object
                    properties:
                      secret:
                        type: object
                        description: Docker swarm.Secret object (data always nil)
                      services:
                        type: array
                        items:
                          $ref: "#/components/schemas/ServiceRef"
        "304":
          $ref: "#/components/responses/NotModified"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: removeSecret
      tags: [Secrets]
      summary: Remove a secret
      description: Removes a secret from the swarm. Fails if the secret is in use by a service. Requires operations level 3.
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "204":
          description: Secret removed
        "403":
          description: Operations level too low
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /search:
    get:
      operationId: search
      tags: [Search]
      summary: Global cross-resource search
      description: |
        Searches names, images, and labels across all resource types.
        Results are grouped by type. The `limit` parameter controls max results
        per type (default 3; 0 means up to 1000 per type).
      parameters:
        - name: q
          in: query
          required: true
          description: Search query (case-insensitive substring match, max 200 chars)
          schema:
            type: string
            maxLength: 200
        - name: limit
          in: query
          description: Max results per resource type (0 = up to 1000)
          schema:
            type: integer
            default: 3
            minimum: 0
      responses:
        "200":
          description: Grouped search results
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "304":
          $ref: "#/components/responses/NotModified"
        "400":
          $ref: "#/components/responses/BadRequest"

  # ── Topology ─────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-cetacean_session
      description: Session cookie set after OIDC authentication
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token for machine-to-machine OIDC authentication

  # ── Parameters ───────────────────────────────────────────────────────
  parameters:
    resourceId:
      name: id
      in: path
      required: true
      description: Resource ID
      schema:
        type: string
    search:
      name: search
      in: query
      description: Case-insensitive substring search on resource names
      schema:
        type: string
    filter:
      name: filter
      in: query
      description: |
        expr-lang filter expression (max 512 characters).
        Available fields depend on the resource type.
      schema:
        type: string
        maxLength: 512
    limit:
      name: limit
      in: query
      description: Page size
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    offset:
      name: offset
      in: query
      description: Starting offset for pagination
      schema:
        type: integer
        minimum: 0
        default: 0
    dir:
      name: dir
      in: query
      description: Sort direction
      schema:
        type: string
        enum: [asc, desc]
        default: asc
...
  headers:
    ETag:
      description: Weak ETag for conditional requests (use with If-None-Match)
      schema:
        type: string
        example: '"w/abc123"'
    Link:
      description: RFC 8288 pagination links (rel="next", rel="prev")
      schema:
        type: string
        example: '</nodes?limit=50&offset=50>; rel="next"'
...
  responses:
    NotModified:
      description: Not modified (ETag matched If-None-Match)
    BadRequest:
      description: Bad request (invalid filter, missing parameter, etc.)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
    Conflict:
      description: Version conflict — the resource was modified concurrently; retry the operation
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
  schemas:
...
    ProblemDetail:
      type: object
      description: |
        RFC 9457 Problem Details with JSON-LD @context.
        Domain-specific errors use `/api/errors/{code}` as the type URI
        (e.g. `/api/errors/SVC001`). Generic errors use `about:blank`.
      required: [type, title, status]
      properties:
        "@context":
          type: string
          example: /api/context.jsonld
        type:
          type: string
          description: "Problem type URI — `/api/errors/{code}` for domain errors, `about:blank` for generic"
          example: /api/errors/SVC001
        title:
          type: string
          description: Short human-readable summary
          example: Service Version Conflict
        status:
          type: integer
          description: HTTP status code
          example: 409
        detail:
          type: string
          description: Human-readable explanation
        instance:
          type: string
          description: Request path that generated the error
        requestId:
          type: string
          description: Correlation ID for this request
...
    DetailEnvelope:
      type: object
      description: JSON-LD wrapper for detail endpoints
      properties:
        "@context":
          type: string
          example: /api/context.jsonld
        "@id":
          type: string
          description: Canonical path of this resource
          example: /nodes/abc123
        "@type":
          type: string
          description: Resource type name
          example: Node
    CollectionEnvelope:
      type: object
      description: JSON-LD wrapper for list endpoints
      required: [items, total, limit, offset]
      properties:
        "@context":
          type: string
          example: /api/context.jsonld
        "@type":
          type: string
          example: Collection
        items:
          type: array
          items: {}
        total:
          type: integer
          description: Total number of items matching the query
        limit:
          type: integer
          description: Page size used
        offset:
          type: integer
          description: Starting offset used

    # ── Collections ──────────────────────────────────────────────────
    NodeCollection:
      allOf:
        - $ref: "#/components/schemas/CollectionEnvelope"
        - type: object
          properties:
            items:
              type: array
              items:
                type: object
                description: Docker swarm.Node object
    ServiceCollection:
      allOf:
        - $ref: "#/components/schemas/CollectionEnvelope"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ServiceListItem"
    TaskCollection:
      allOf:
        - $ref: "#/components/schemas/CollectionEnvelope"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/EnrichedTask"
...
    SecretCollection:
      allOf:
        - $ref: "#/components/schemas/CollectionEnvelope"
        - type: object
          properties:
            items:
              type: array
              items:
                type: object
                description: Docker swarm.Secret object (data always nil)
...
    ServiceRef:
      type: object
      description: Lightweight service cross-reference
      properties:
        "@id":
          type: string
          description: Canonical path
          example: /services/abc123
        id:
          type: string
        name:
          type: string
    ServiceListItem:
      type: object
      description: Service with running task count (extends Docker swarm.Service)
      properties:
        ID:
          type: string
        Spec:
          type: object
          description: Docker ServiceSpec
        RunningTasks:
          type: integer
          description: Number of currently running tasks
    EnrichedTask:
      type: object
      description: Docker swarm.Task with added ServiceName and NodeHostname
      properties:
        ID:
          type: string
        ServiceID:
          type: string
        NodeID:
          type: string
        Status:
          type: object
          description: Docker TaskStatus
        ServiceName:
          type: string
          description: Resolved from cache
        NodeHostname:
          type: string
          description: Resolved from cache
...
    SearchResponse:
      type: object
      properties:
        "@context":
          type: string
          example: /api/context.jsonld
        "@id":
          type: string
          example: /search
        "@type":
          type: string
          example: SearchResult
        query:
          type: string
        results:
          type: object
          description: Grouped results keyed by resource type
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/SearchResult"
        counts:
          type: object
          description: Total match count per type (may exceed returned results)
          additionalProperties:
            type: integer
        total:
          type: integer
          description: Total matches across all types
    SearchResult:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        detail:
          type: string
          description: Context-dependent detail (image, driver, date, etc.)
        state:
          type: string
          description: Present for services and tasks

    # ── Topology ─────────────────────────────────────────────────────
...
    Integration:
      description: A detected third-party integration on a service
      discriminator:
        propertyName: name
        mapping:
          traefik: "#/components/schemas/TraefikIntegration"
          shepherd: "#/components/schemas/ShepherdIntegration"
          swarm-cronjob: "#/components/schemas/CronjobIntegration"
          diun: "#/components/schemas/DiunIntegration"
      oneOf:
        - $ref: "#/components/schemas/TraefikIntegration"
        - $ref: "#/components/schemas/ShepherdIntegration"
        - $ref: "#/components/schemas/CronjobIntegration"
        - $ref: "#/components/schemas/DiunIntegration"
    TraefikIntegration:
      type: object
      description: Traefik reverse proxy integration detected from service labels
      required: [name]
      properties:
        name:
          type: string
          enum: [traefik]
          description: Discriminator field identifying the integration type
        enabled:
          type: boolean
          description: Whether the Traefik integration is enabled
        routers:
          type: array
          items:
            $ref: "#/components/schemas/TraefikRouter"
        services:
          type: array
          items:
            $ref: "#/components/schemas/TraefikService"
        middlewares:
          type: array
          items:
            $ref: "#/components/schemas/TraefikMiddleware"
    TraefikRouter:
      type: object
      description: A Traefik router detected from service labels
      properties:
        name:
          type: string
          description: Router name
        rule:
          type: string
          description: Routing rule expression (e.g. "Host(`example.com`)")
        entrypoints:
          type: array
          items:
            type: string
          description: Entrypoint names this router listens on
        tls:
          $ref: "#/components/schemas/TraefikTLS"
        middlewares:
          type: array
          items:
            type: string
          description: Middleware names attached to this router
        service:
          type: string
          description: Backend service name this router forwards to
        priority:
          type: integer
          description: Router priority (higher wins on rule conflicts)
    TraefikTLS:
      type: object
      description: TLS configuration for a Traefik router
      properties:
        certResolver:
          type: string
          description: Certificate resolver name (e.g. "letsencrypt")
        domains:
          type: array
          items:
            $ref: "#/components/schemas/TraefikTLSDomain"
          description: Explicit domain entries for certificate generation
        options:
          type: string
          description: TLS options name
    TraefikTLSDomain:
      type: object
      description: A domain entry in Traefik TLS configuration
      properties:
        main:
          type: string
          description: Primary domain name
        sans:
          type: array
          items:
            type: string
          description: Subject alternative names
    TraefikService:
      type: object
      description: A Traefik service detected from service labels
      properties:
        name:
          type: string
          description: Service name
        port:
          type: integer
          description: Backend port number
        scheme:
          type: string
          description: Backend scheme (e.g. "http", "https", "h2c")
    TraefikMiddleware:
      type: object
      description: A Traefik middleware detected from service labels
      properties:
        name:
          type: string
          description: Middleware name
        type:
          type: string
          description: Middleware type (e.g. "redirectScheme", "basicAuth")
        config:
          type: object
          additionalProperties: true
          description: Middleware configuration key-value pairs
    ShepherdIntegration:
      type: object
      description: Shepherd auto-update integration detected from service labels
      required: [name]
      properties:
        name:
          type: string
          enum: [shepherd]
        enabled:
          type: boolean
        authConfig:
          type: string
          description: Registry auth configuration ID
    CronjobIntegration:
      type: object
      description: Swarm Cronjob integration detected from service labels
      required: [name]
      properties:
        name:
          type: string
          enum: [swarm-cronjob]
        enabled:
          type: boolean
        schedule:
          type: string
          description: Cron expression
        skipRunning:
          type: boolean
        replicas:
          type: integer
        registryAuth:
          type: boolean
        queryRegistry:
          type: boolean
    DiunIntegration:
      type: object
      description: Diun image update notifier integration detected from service labels
      required: [name]
      properties:
        name:
          type: string
          enum: [diun]
        enabled:
          type: boolean
        regopt:
          type: string
          description: Registry options name
        watchRepo:
          type: boolean
        notifyOn:
          type: string
          description: "Semicolon-separated triggers (new, update)"
        sortTags:
          type: string
          enum: [default, reverse, semver, lexicographical]
        maxTags:
          type: integer
        includeTags:
          type: string
          description: Regex for tag inclusion
        excludeTags:
          type: string
          description: Regex for tag exclusion
        hubLink:
          type: string
        platform:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string